move custom provider example to example/main.go
